# docs/openapi.json
{
  "components": {
    "schemas": {
      "FeedbackBody": {
        "additionalProperties": false,
        "description": "``POST /jobs/{id}/feedback``'s body (api.md \u00a76).\n\n``source`` is honoured only when neither a token nor ``X-Client-Name`` names the caller.",
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "boolean"
          },
          "edited": {
            "default": false,
            "title": "Edited",
            "type": "boolean"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "quality_score": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Score"
          },
          "source": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "validation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FeedbackValidationBody"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "accepted"
        ],
        "title": "FeedbackBody",
        "type": "object"
      },
      "FeedbackValidationBody": {
        "additionalProperties": false,
        "description": "The caller's own validation verdict (api.md \u00a76's ``validation`` object).",
        "properties": {
          "detail": {
            "title": "Detail"
          },
          "passed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Passed"
          }
        },
        "title": "FeedbackValidationBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobBody": {
        "additionalProperties": false,
        "description": "``POST /jobs``'s body: ``/generate``'s plus class, priority, wait bound and idempotency.",
        "properties": {
          "class": {
            "default": "normal",
            "pattern": "^(interactive|normal|background|batch)$",
            "title": "Class",
            "type": "string"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "idempotent": {
            "default": true,
            "title": "Idempotent",
            "type": "boolean"
          },
          "max_wait_seconds": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Wait Seconds"
          },
          "messages": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/MessageBody"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Messages"
          },
          "overrides": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OverridesBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "priority": {
            "anyOf": [
              {
                "maximum": 999.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt"
          },
          "response_format": {
            "anyOf": [
              {
                "pattern": "^(text|json|json_schema)$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Format"
          },
          "sampling": {
            "additionalProperties": true,
            "title": "Sampling",
            "type": "object"
          },
          "stream": {
            "default": false,
            "title": "Stream",
            "type": "boolean"
          },
          "system": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "System"
          },
          "task": {
            "title": "Task",
            "type": "string"
          }
        },
        "required": [
          "task"
        ],
        "title": "JobBody",
        "type": "object"
      },
      "MessageBody": {
        "additionalProperties": false,
        "description": "One turn of a caller-supplied transcript (api.md \u00a74).",
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "role": {
            "pattern": "^(system|user|assistant|tool)$",
            "title": "Role",
            "type": "string"
          },
          "tool_call_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Call Id"
          }
        },
        "required": [
          "role",
          "content"
        ],
        "title": "MessageBody",
        "type": "object"
      },
      "OverridesBody": {
        "additionalProperties": false,
        "description": "Routing \u00a710's overrides, as far as this phase implements them.",
        "properties": {
          "disallow_fallback": {
            "default": false,
            "title": "Disallow Fallback",
            "type": "boolean"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "require_evidence": {
            "default": false,
            "title": "Require Evidence",
            "type": "boolean"
          },
          "runtime_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuntimeProfileOverrideBody"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "OverridesBody",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "LoadCoach",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/jobs": {
      "get": {
        "description": "Jobs newest first, filtered by state, class, task and source; cursor-paginated.",
        "operationId": "get_jobs_api_v1_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "class",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Class"
            }
          },
          {
            "in": "query",
            "name": "task",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Task"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List jobs",
        "tags": [
          "jobs"
        ]
      },
      "post": {
        "description": "Enqueue ``body`` and return the job. ``202``; a repeated key returns the original job.\n\nErrors: ``TASK_PROFILE_NOT_FOUND``, ``VALIDATION_ERROR`` (a priority outside the class's\nband), ``QUEUE_FULL`` (429).",
        "operationId": "post_job_api_v1_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Post Job Api V1 Jobs Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit a job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "description": "The job: state, attempts, routing summary, usage, timings, validation, degradations.",
        "operationId": "get_job_document_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Job Document Api V1 Jobs  Job Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One job in full",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/cancel": {
      "post": {
        "description": "Cancel: at once for a waiting job, at the next chunk boundary for an executing one.\n\n``409 JOB_NOT_CANCELLABLE`` for a terminal job; idempotent otherwise.",
        "operationId": "post_cancel_api_v1_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Post Cancel Api V1 Jobs  Job Id  Cancel Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel a job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/explanation": {
      "get": {
        "description": "The routing decision whose ``job_id`` matches \u2014 a lookup, never a copy (LCX3).",
        "operationId": "get_job_explanation_api_v1_jobs__job_id__explanation_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Job Explanation Api V1 Jobs  Job Id  Explanation Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The job's routing explanation",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/feedback": {
      "post": {
        "description": "Record the caller's verdict on the job; ``write`` scope (spec \u00a714).\n\n``201`` with the stored record on a source's first feedback for the job, ``200`` on an\nupdate; idempotent per ``(job_id, source)``. ``404 JOB_NOT_FOUND`` otherwise.",
        "operationId": "post_feedback_api_v1_jobs__job_id__feedback_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Post Feedback Api V1 Jobs  Job Id  Feedback Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Caller feedback on a job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/stream": {
      "get": {
        "description": "SSE: state changes, tokens when streaming was requested, and the terminal event.\n\nReplays the persisted events after ``Last-Event-ID`` and follows the live broker; closes on\nthe terminal event.",
        "operationId": "get_job_stream_api_v1_jobs__job_id__stream_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The job's event stream",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}